auto 0 24 191

+++ test/test_pat.c
@@ -12,9 +12,9 @@ int main(int argc, char *argv[])
 	struct pat_captures cap = pat_capture(plain, ferrets, 0);
 	size_t i = 0;
 	for(;i<cap.ovecsize;++i) {
-		printf("capture %ld %ld %.*s",
+		printf("capture %ld %d %.*s",
 			   i, cap.ovector[i],
-			   LITSIZ("ferrets"),
+			   (int)LITSIZ("ferrets"),
 			   ferrets.base + cap.ovector[i]);
 	}
